Fixed catching unnececcary exceptions
Exception Abort can't be thrown, so it is useless.

Catching SystemExit and KeyboardInterrupt makes little sense as the thread can
be terminated by KeyboardInterrupt before or after entering the wrapped command
processing function. So one have to handle these exceptions in his main function
anyway.

Catching any exception in order to print a message and re-raise it changes the
standard behaviour and might seem confusing.